// src/models/enums.ts
export enum OrderTypeCategory {
  Veg = 'veg',
  NonVeg = 'non-veg',
}

export enum OrderSizeCategory {
  Small = 'small',
  Medium = 'medium',
  Large = 'large',
}

export enum OrderBaseCategory {
  Thin = 'thin',
  Thick = 'thick',
}

export enum OrderStages {
  OrderPlaced = 'Order Placed',
  OrderInMaking = 'Order in Making',
  OrderReady = 'Order Ready',
  OrderPicked = 'Order Picked',
}